+++ luxurydelivery/src/main/java/com/deliverysl/luxurydelivery/product/dto/FoodDTO.java
@@ -2,13 +2,14 @@ package com.deliverysl.luxurydelivery.product.dto;
 
 import com.deliverysl.luxurydelivery.allergen.dto.ShortAllergenDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record FoodDTO(
         Long id,
         String name,
         String image,
-        double price,
+        BigDecimal price,
         String category,
         int calories,
         List<ShortAllergenDTO> allergenList
